requirements.txt
# requirements.txt
pyyaml